+++ src/s21_graph_algorithms/tests/test_dijkstras.cc
@@ -3,7 +3,7 @@
 #include "../s21_graph_algorithms.h"
 
 TEST(dijkstra, test_1) {
-  int graph6x6[] = {0, 7, 9, 0, 0, 14,
+  double graph6x6[] = {0, 7, 9, 0, 0, 14,
                     7, 0, 10, 15, 0, 0,
                     9, 10, 0, 11, 0, 2,
                     0,  15, 11, 0, 6, 0,
@@ -11,12 +11,12 @@ TEST(dijkstra, test_1) {
                     14, 0, 2, 0, 9, 0};
   s21::Graph graph(graph6x6, 6);
   s21::GraphAlgorithms graph_algorithms;
-  int result = graph_algorithms.getShortestPathBetweenVertices(graph, 1, 6);
+  double result = graph_algorithms.getShortestPathBetweenVertices(graph, 1, 6);
   ASSERT_EQ(result, 11);
 }
 
 TEST(dijkstra, test_2) {
-  int graph9x9[] = {0, 4, 0, 0, 0, 0, 0, 8, 0,
+  double graph9x9[] = {0, 4, 0, 0, 0, 0, 0, 8, 0,
                     4, 0, 8, 0, 0, 0, 0, 11, 0,
                     0, 8, 0, 7, 0, 4, 0, 0, 2,
                     0, 0, 7, 0, 9, 14, 0, 0, 0,
@@ -27,6 +27,6 @@ TEST(dijkstra, test_2) {
                     0, 0, 2, 0, 0, 0, 6, 7, 0};
   s21::Graph graph(graph9x9, 9);
   s21::GraphAlgorithms graph_algorithms;
-  int result = graph_algorithms.getShortestPathBetweenVertices(graph, 1, 5);
+  double result = graph_algorithms.getShortestPathBetweenVertices(graph, 1, 5);
   ASSERT_EQ(result, 21);
 }
